Changes: airport apis added

+++ flight-service/src/main/java/com/airline/flightservice/repository/AirportRepository.java
@@ -2,4 +2,4 @@ package com.airline.flightservice.repository;
 
 import com.airline.flightservice.entity.Airport;
 
-public interface AirportRepository extends BaseRepository<Airport, Integer>{}
+public interface AirportRepository extends BaseRepository<Airport, Long>{}
